+++ InspectCoroutine.lua
@@ -1,6 +1,8 @@
 local addonName, miog = ...
 local wticc = WrapTextInColorCode
 
+local partyPool
+
 local groupSystem = {}
 groupSystem.groupMember = {}
 
@@ -14,7 +16,6 @@ MIOG_SavedSpecIDs = {}
 local timers = {}
 
 local currentInspection = nil
-local currentInspectionName = nil
 
 local function createGroupMemberEntry(guid, unitID)
 	local _, classFile, _, _, _, name = GetPlayerInfoByGUID(guid)
@@ -33,12 +34,14 @@ local function createGroupMemberEntry(guid, unitID)
 
 end
 
-local function updateRosterInfoData()	
+local function updateRosterInfoData()
 	if(currentInspection and GetTimePreciseSec() - lastNotifyTime > 7) then
 		ClearInspectPlayer(true)
 		updateRosterInfoData()
 
 	end
+
+	partyPool:ReleaseAll()
 
 	groupSystem.groupMember = {}
 
@@ -89,6 +92,7 @@ local function updateRosterInfoData()
 					name = name,
 					shortName = UnitName(unitID),
 					classID = fileName and miog.CLASSFILE_TO_ID[fileName],
+					classFileName = fileName,
 					role = combatRole,
 				}
 
@@ -96,7 +100,6 @@ local function updateRosterInfoData()
 
 					if(currentInspection == nil and UnitIsConnected(unitID) and CanInspect(unitID)) then
 
-						currentInspectionName = name
 						currentInspection = guid
 
 						if(timers[guid]) then
@@ -118,7 +121,6 @@ local function updateRosterInfoData()
 						function()
 							if(groupSystem.groupMember[guid]) then
 								lastNotifyTime = GetTimePreciseSec()
-			
 
 								NotifyInspect(unitID) -- 2nd argument for hook to check if own or other addons' notify
 			
@@ -137,8 +139,8 @@ local function updateRosterInfoData()
 					end
 				end
 
-				if(groupSystem.classCount[miog.CLASSFILE_TO_ID[fileName]]) then
-					groupSystem.classCount[miog.CLASSFILE_TO_ID[fileName]] = groupSystem.classCount[miog.CLASSFILE_TO_ID[fileName]] + 1
+				if(groupSystem.classCount[miog.CLASSFILE_TO_ID[fileName] ]) then
+					groupSystem.classCount[miog.CLASSFILE_TO_ID[fileName] ] = groupSystem.classCount[miog.CLASSFILE_TO_ID[fileName] ] + 1
 
 				end
 	
@@ -171,6 +173,10 @@ local function updateRosterInfoData()
 			local specID, _, _, _, role = GetSpecializationInfo(GetSpecialization())
 			member.specID = specID
 			member.role = role
+		
+		elseif(miog.checkSystem.groupMember[member.guid]) then
+			MIOG_SavedSpecIDs[guid] = miog.checkSystem.groupMember[member.guid].specId ~= 0 and miog.checkSystem.groupMember[member.guid].specId or nil
+			MIOG_InspectedGUIDs[guid] = MIOG_SavedSpecIDs[guid] and GetTimePreciseSec() or nil
 		
 		end
 
@@ -257,27 +263,27 @@ local function updateRosterInfoData()
 
 	miog.ClassPanel.Container:MarkDirty()
 
-	if(numOfMembers < 5) then
+	--[[if(numOfMembers < 5) then
 		if(groupSystem.roleCount["TANK"] < 1) then
 			indexedGroup[#indexedGroup + 1] = {guid = "emptyTank", unitID = "emptyTank", name = "afkTank", classID = 20, role = "TANK", icon = miog.C.STANDARD_FILE_PATH .. "/infoIcons/empty.png"}
-			--groupSystem.roleCount["TANK"] = groupSystem.roleCount["TANK"] + 1
+			groupSystem.roleCount["TANK"] = groupSystem.roleCount["TANK"] + 1
 		end
 
 		if(groupSystem.roleCount["HEALER"] < 1 and #indexedGroup < 5) then
 			indexedGroup[#indexedGroup + 1] = {guid = "emptyHealer", unitID = "emptyHealer", name = "afkHealer", classID = 21, role = "HEALER", icon = miog.C.STANDARD_FILE_PATH .. "/infoIcons/empty.png"}
-			--groupSystem.roleCount["HEALER"] = groupSystem.roleCount["HEALER"] + 1
+			groupSystem.roleCount["HEALER"] = groupSystem.roleCount["HEALER"] + 1
 
 		end
 
 		for i = 1, 3 - groupSystem.roleCount["DAMAGER"], 1 do
 			if(groupSystem.roleCount["DAMAGER"] < 3 and #indexedGroup < 5) then
 				indexedGroup[#indexedGroup + 1] = {guid = "emptyDPS" .. i, unitID = "emptyDPS" .. i, name = "afkDPS" .. i, classID = 22, role = "DAMAGER", icon = miog.C.STANDARD_FILE_PATH .. "/infoIcons/empty.png"}
-				--groupSystem.roleCount["DAMAGER"] = groupSystem.roleCount["DAMAGER"] + 1
+				groupSystem.roleCount["DAMAGER"] = groupSystem.roleCount["DAMAGER"] + 1
 
 			end
 
 		end
-	end
+	end]]
 
 	table.sort(indexedGroup, function(k1, k2)
 		if(k1.role ~= k2.role) then
@@ -290,8 +296,76 @@ local function updateRosterInfoData()
 	end)
 
 	miog.ApplicationViewer.TitleBar.GroupComposition.Roles:SetText(groupSystem.roleCount["TANK"] .. "/" .. groupSystem.roleCount["HEALER"] .. "/" .. groupSystem.roleCount["DAMAGER"])
+
+	for index, member in ipairs(indexedGroup) do
+		local memberFrame = partyPool:Acquire()
+		memberFrame:SetWidth(miog.PartyCheck:GetWidth())
+		memberFrame.Name:SetText(WrapTextInColorCode(member.shortName, C_ClassColor.GetClassColor(member.classFileName):GenerateHexColor()))
+		memberFrame.Role:SetTexture(miog.C.STANDARD_FILE_PATH .."/infoIcons/" .. member.role .. "Icon.png")
+		memberFrame.Spec:SetTexture(miog.SPECIALIZATIONS[member.specID].squaredIcon)
+
+		if(miog.checkSystem.groupMember[member.guid]) then
+			local checkData = miog.checkSystem.groupMember[member.guid]
+			memberFrame.ILvl:SetText(checkData.ilvl or "N/A")
+			memberFrame.Durability:SetText(checkData.durability or "N/A")
+
+			for k, v in pairs(memberFrame.MissingEnchantSlots) do
+				if(type(v) == "table") then
+					if(checkData.missingEnchants and checkData.missingEnchants[k]) then
+						v.layoutIndex = k
+						v:Show()
+
+					else
+						v.layoutIndex = nil
+						v:Hide()
+
+					end
+				end
+
+			end
+
+			memberFrame.MissingEnchantSlots:MarkDirty()
+
+			local keystoneInfo = miog.checkSystem.keystoneData[member.name]
+
+			if(keystoneInfo) then
+				local mapName, id, timeLimit, texture, background = C_ChallengeMode.GetMapUIInfo(keystoneInfo.challengeMapID)
+
+				memberFrame.Keystone:SetTexture(texture)
+				memberFrame.Keylevel:SetText("+" .. keystoneInfo.level)
+			end
+		end
 		
+
+		local rioProfile = RaiderIO.GetProfile(member.name)
+
+		if(rioProfile) then
+			memberFrame.Score:SetText(rioProfile.mythicKeystoneProfile and rioProfile.mythicKeystoneProfile.currentScore or "N/A")
+			memberFrame.Progress:SetText("")
+			
+			local currentData, nonCurrentData, orderedData = miog.getRaidSortData(member.name)
+			
+			for a, b in ipairs(orderedData) do
+				memberFrame.Progress:SetText((memberFrame.Progress:GetText() or "") .. b.parsedString .. " ")
+
+				if(a == 3) then
+					break
+				end
+			end
+		else
+			memberFrame.Score:SetText("N/A")
+			memberFrame.Progress:SetText("N/A")
+		
+		end
+
+		memberFrame.layoutIndex = member.role == "TANK" and 1 or member.role == "HEALER" and 50 or 100
+		memberFrame:Show()
+	end
+
+	miog.PartyCheck.ScrollFrame.Container:MarkDirty()
 end
+
+miog.updateRosterInfoData = updateRosterInfoData
 
 local function resetInspectCoroutine()
 	MIOG_InspectedGUIDs = {}
@@ -376,12 +450,22 @@ end)
 hooksecurefunc("ClearInspectPlayer", function(own)
 	if(own) then
 		currentInspection = nil
-		currentInspectionName = nil
 		--miog.pveFrame2.ClassPanelStatusString:SetText("Waiting for next inspect...")
 	end
 end)
 
+
+local function resetPartyFrames(_, frame)
+    frame:Hide()
+    frame.layoutIndex = nil
+
+    frame.Name:SetText("")
+    frame.Role:SetTexture(nil)
+    frame.Spec:SetTexture(nil)
+end
+
 miog.createInspectCoroutine = function()
+	--miog.OnUnitUpdate()
 	eventReceiver:RegisterEvent("PLAYER_ENTERING_WORLD")
 	eventReceiver:RegisterEvent("GROUP_JOINED")
 	eventReceiver:RegisterEvent("GROUP_LEFT")
@@ -389,5 +473,8 @@ miog.createInspectCoroutine = function()
 	eventReceiver:RegisterEvent("INSPECT_READY")
 	eventReceiver:RegisterEvent("GROUP_ROSTER_UPDATE")
 	eventReceiver:SetScript("OnEvent", inspectCoroutineEvents)
-end
 
+    partyPool = CreateFramePool("Frame", miog.PartyCheck.ScrollFrame.Container, "MIOG_PartyCheckPlayerTemplate", resetPartyFrames)
+
+	--local allUnitsInfo = miog.openRaidLib.GetAllUnitsInfo()
+end